Update Decision Environments Pages to use PageRoutedTabs. Add Back to Decision Environments tab.
