+++ frontend/eda/Resources/decision-environments/DecisionEnvironmentPage/DecisionEnvironmentPage.tsx
@@ -1,0 +1,100 @@
+import { ButtonVariant } from '@patternfly/react-core';
+import { DropdownPosition } from '@patternfly/react-core/deprecated';
+import { PencilAltIcon, TrashIcon } from '@patternfly/react-icons';
+import { useMemo } from 'react';
+import { useTranslation } from 'react-i18next';
+import { useParams } from 'react-router-dom';
+import {
+  IPageAction,
+  PageActionSelection,
+  PageActionType,
+  PageActions,
+  PageHeader,
+  PageLayout,
+  useGetPageUrl,
+  usePageNavigate,
+} from '../../../../../framework';
+import { useGet } from '../../../../common/crud/useGet';
+import { EdaRoute } from '../../../EdaRoutes';
+import { API_PREFIX, SWR_REFRESH_INTERVAL } from '../../../constants';
+import { EdaDecisionEnvironmentRead } from '../../../interfaces/EdaDecisionEnvironment';
+import { useDeleteDecisionEnvironment } from '../hooks/useDeleteDecisionEnvironments';
+import { PageRoutedTabs } from '../../../../../framework/PageTabs/PageRoutedTabs';
+
+export function DecisionEnvironmentPage() {
+  const { t } = useTranslation();
+  const params = useParams<{ id: string }>();
+  const pageNavigate = usePageNavigate();
+
+  const { data: decisionEnvironment } = useGet<EdaDecisionEnvironmentRead>(
+    `${API_PREFIX}/decision-environments/${params.id ?? ''}/`,
+    undefined,
+    { refreshInterval: SWR_REFRESH_INTERVAL }
+  );
+
+  const deleteDecisionEnvironment = useDeleteDecisionEnvironment((deleted) => {
+    if (deleted.length > 0) {
+      pageNavigate(EdaRoute.DecisionEnvironments);
+    }
+  });
+
+  const itemActions = useMemo<IPageAction<EdaDecisionEnvironmentRead>[]>(
+    () => [
+      {
+        type: PageActionType.Button,
+        variant: ButtonVariant.primary,
+        selection: PageActionSelection.Single,
+        icon: PencilAltIcon,
+        isPinned: true,
+        label: t('Edit decision environment'),
+        onClick: (decisionEnvironment: EdaDecisionEnvironmentRead) =>
+          pageNavigate(EdaRoute.EditDecisionEnvironment, {
+            params: { id: decisionEnvironment.id },
+          }),
+      },
+      {
+        type: PageActionType.Seperator,
+      },
+      {
+        type: PageActionType.Button,
+        selection: PageActionSelection.Single,
+        icon: TrashIcon,
+        label: t('Delete decision environment'),
+        onClick: (decisionEnvironment: EdaDecisionEnvironmentRead) =>
+          deleteDecisionEnvironment([decisionEnvironment]),
+        isDanger: true,
+      },
+    ],
+    [deleteDecisionEnvironment, pageNavigate, t]
+  );
+
+  const getPageUrl = useGetPageUrl();
+
+  return (
+    <PageLayout>
+      <PageHeader
+        title={decisionEnvironment?.name}
+        breadcrumbs={[
+          { label: t('Decision Environments'), to: getPageUrl(EdaRoute.DecisionEnvironments) },
+          { label: decisionEnvironment?.name },
+        ]}
+        headerActions={
+          <PageActions<EdaDecisionEnvironmentRead>
+            actions={itemActions}
+            position={DropdownPosition.right}
+            selectedItem={decisionEnvironment}
+          />
+        }
+      />
+      <PageRoutedTabs
+        backTab={{
+          label: t('Back to Decision Environments'),
+          page: EdaRoute.DecisionEnvironments,
+          persistentFilterKey: 'decision_environments',
+        }}
+        tabs={[{ label: t('Details'), page: EdaRoute.DecisionEnvironmentDetails }]}
+        params={{ id: decisionEnvironment?.id }}
+      />
+    </PageLayout>
+  );
+}
